fix(backlog): M11/M12 安全批次 — harness 三态退码/路径参数化 + tsdf widen + DB死链 + dev接线
清 backlog 中低风险、可独立验证、无设计分叉的项:

M12.1 harness 警告档吞退码 → 三态(FAIL→exit1):
- vin_restore(假重合 NCC<0.55 从 WARN 降 FAIL)、eys3d_mode25(坏 scale 深度物理不合理升 FAIL)、
  worker_vin_pipeline(run.sh 此前根本不调 analyze、不 gate → 末尾真调 + exit $?)、
  vehicle_axle(真值缺失从静默 exit0 → loud-fail exit1)、berxel_mix_replay/cv_engine_smoke 复核已三态。
M12.4 机器特定绝对路径 → 环境变量 + 默认探测(default=原值,本机行为不变,缺则 loud-fail):
- cv_models_smoke/cv_modelregistry_load/cv_vin_pipeline/cv_yolo_detect 模型路径、eys3d_depth_test.cpp ZD 表路径、
  vehicle_axle 真值路径(JCHY_DATA + 多候选探测);depth_* 复核已是 env 可覆盖,无需改。
M12.3(cv_yolo)永真断言 → 真判据:count>=0 / count2>=count1 换成"空白图必 0 框 + 低阈仍 0 框"(stub 会被识破)。
M12.7 文案/阈值漂移订正:vin_restore(NCC>0.6→0.55)、scan_multiview_quality(coverage 门 78/92→88/94)。
M12.8 logs_upload 缺 must 场景静默判过 → must 清单缺失 exit1;vin_restore 产物支持 OUTPUT_DIR 覆盖。
M11.16 TsdfStats allocated/integrated_voxels int → int64_t(消除 grid_dim>1290 回写截断,去 static_cast<int>)。
M11.8 删 MIGRATION_1_2 死链(v1 从未发布、相对 v2 不完整)+ 从 DatabaseModule.addMigrations 摘除。
M11.3 devserver 启动调 auth.EnableBearerDirect(true)(仅 dev,生产单一信任源);
M11.4 laserworker 建 pool 后 SetAuditRecorder(audit.NewPG(pool))(标定级操作审计真落 PG)。

验证:py_compile + bash -n 全过;go build/vet/test 净;./dev.sh build 成功;native host 5 runner 全 [OK]。

+++ native/reconstruction/tsdf.cpp
@@ -111,13 +111,11 @@ void TsdfVolume::Get(int i, int j, int k, float& out_sdf, float& out_weight) con
 
 TsdfStats TsdfVolume::Stats() const {
     TsdfStats s{};
-    // grid_dim^3 大网格（如 200^3=8M、512^3=134M）整型乘法会 int 溢出（UB），先用 size_t 算。
-    // TODO(tsdf-stats): TsdfStats::allocated_voxels 字段本身仍是 int（在 tsdf.h，本轮不改 header
-    //   避免跨文件结构改动），grid_dim>1290 时回写 int 仍会截断；终态应把该字段及 integrated_voxels
-    //   widen 到 int64_t/size_t。当前真机 grid_dim≤512（134M）已超 int 上限(2.1G)前的安全区但临界，先消除乘法 UB。
+    // grid_dim^3 大网格（如 200^3=8M、512^3=134M）整型乘法会溢出（UB），用 size_t 算；
+    // TsdfStats::allocated_voxels/integrated_voxels 已 widen 到 int64_t（tsdf.h），回写不截断。
     const std::size_t total_voxels =
         static_cast<std::size_t>(grid_dim_) * grid_dim_ * grid_dim_;
-    s.allocated_voxels = static_cast<int>(total_voxels);
+    s.allocated_voxels = static_cast<int64_t>(total_voxels);
     s.integrated_voxels = 0;
     for (float w : weight_) if (w > 0.0f) ++s.integrated_voxels;
     s.integrated_frames = frame_count_;
